requirements.txt
django
django-apscheduler
django-cors-headers
requests